+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- [[#14](https://github.com/schemadb/node-sdk/issues/14)] Get specific schema version.
+
 ## [0.2.0] - 2020-04-15
 
 ### Added